refactored so state and handlers are passed as props
