+++ front-end/src/forms/ReservationForm.js
@@ -1,67 +1,14 @@
-import React, { useState } from "react";
-import { useHistory } from "react-router-dom";
-import { createReservation } from "../utils/api";
+import React from "react";
 import CancelBtn from "../buttons/CancelBtn";
 import SubmitBtn from "../buttons/SubmitBtn";
-import ErrorList from "../layout/ErrorList";
 
-export default function ReservationForm() {
-  // initial values of input fields
-  const initialFormState = {
-    first_name: "",
-    last_name: "",
-    mobile_number: "",
-    reservation_date: "",
-    reservation_time: "",
-    people: "",
-  };
-
-  /**
-   * state management of form input fields and error
-   */
-  const [formData, setFormData] = useState({ ...initialFormState });
-  const [reservationError, setReservationError] = useState(null);
-
-  const history = useHistory();
-
-  /**
-   * functions to handle input changes
-   */
-  const handleInputChange = ({ target }) => {
-    if (target.id === "people") {
-      setFormData({
-        ...formData,
-        [target.id]: Number(target.value),
-      });
-    } else {
-      setFormData({
-        ...formData,
-        [target.id]: target.value,
-      });
-    }
-  };
-
-  // sends POST request to api with formData
-  async function handleSubmit(event) {
-    event.preventDefault();
-    const abortController = new AbortController();
-    try {
-      await createReservation(formData, abortController.signal);
-      setFormData({ ...initialFormState });
-      history.push(`/dashboard?date=${formData.reservation_date}`);
-    } catch (err) {
-      setReservationError(err);
-    }
-    
-    
-    return () => abortController.abort();
-  }
-
+export default function ReservationForm({
+  handleInputChange,
+  handleSubmit,
+  formData,
+}) {
   return (
     <div>
-      <div>
-        <ErrorList error={reservationError} />
-      </div>
       <form onSubmit={handleSubmit}>
         <fieldset>
           <div className="row">
